feat(submitentityclaim): start fileupload
