+++ src/common/components/DropZone/Loader.styles.ts
@@ -1,0 +1,8 @@
+export const DropZoneStyles = {
+  width: '100%',
+  textAlign: 'center',
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  alignItems: 'center',
+}
