+++ src/common/components/JsonForm/CustomWidgets/FileUpload/FileUpload.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+interface Props {
+  options: any
+  value: string
+  onChange: (value) => void
+}
+
+const FileUpload: React.FunctionComponent<Props> = () => {
+  return <div />
+}
+
+export default FileUpload
